requirements.txt
dagster==1.0.4
apache-airflow~=2.8.1
apache-airflow-providers-amazon
boto3~=1.21.21
bdrc-volume-manifest-builder>=1.2.6
bdrc-DBApps
internetarchive~=3.0.2
SQLAlchemy~=1.4.39
botocore~=1.24.21
requests
pytest
pytest-helpers-namespace